commonDataProjectsWeb/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
psycopg2-binary==2.9.9
gunicorn==21.2.0
whitenoise==6.6.0
python-decouple==3.8
prometheus-client==0.19.0
django-prometheus==2.3.1

# Note: Project-specific requirements (projects/*/requirements.txt) are automatically merged during Docker build

